Convert deleting a todo to use coroutines
We were representing a deleting action as a `Completable` to keep in
with our Rx chain. Since we're moving to coroutines, we really don't
need that intermediate understanding of an asychronous action that
returns no data. So our return type can just be unit, and we've saved a
lot of code!